requirements.txt
flask==3.1.*
anthropic==0.49.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
python-dotenv==1.1.*
